// rhynode-finance/src/generated/prisma/models/Transaction.ts
export type TransactionWhereInput = {
  AND?: Prisma.TransactionWhereInput | Prisma.TransactionWhereInput[]
  OR?: Prisma.TransactionWhereInput[]
  NOT?: Prisma.TransactionWhereInput | Prisma.TransactionWhereInput[]
  id?: Prisma.StringFilter<"Transaction"> | string
  organizationId?: Prisma.StringFilter<"Transaction"> | string
  userId?: Prisma.StringNullableFilter<"Transaction"> | string | null
  type?: Prisma.StringFilter<"Transaction"> | string
  category?: Prisma.StringNullableFilter<"Transaction"> | string | null
  description?: Prisma.StringFilter<"Transaction"> | string
  amount?: Prisma.DecimalFilter<"Transaction"> | runtime.Decimal | runtime.DecimalJsLike | number | string
  currency?: Prisma.StringFilter<"Transaction"> | string
  date?: Prisma.DateTimeFilter<"Transaction"> | Date | string
  reference?: Prisma.StringNullableFilter<"Transaction"> | string | null
  bankAccountId?: Prisma.StringNullableFilter<"Transaction"> | string | null
  invoiceId?: Prisma.StringNullableFilter<"Transaction"> | string | null
  accountId?: Prisma.StringNullableFilter<"Transaction"> | string | null
  categoryId?: Prisma.StringNullableFilter<"Transaction"> | string | null
  isRecurring?: Prisma.BoolFilter<"Transaction"> | boolean
  receiptId?: Prisma.StringNullableFilter<"Transaction"> | string | null
  scope?: Prisma.StringFilter<"Transaction"> | string
  paymentMethod?: Prisma.StringNullableFilter<"Transaction"> | string | null
  location?: Prisma.StringNullableFilter<"Transaction"> | string | null
  tags?: Prisma.JsonNullableFilter<"Transaction">
  metadata?: Prisma.JsonNullableFilter<"Transaction">
  aiCategory?: Prisma.StringNullableFilter<"Transaction"> | string | null
  aiConfidence?: Prisma.FloatNullableFilter<"Transaction"> | number | null
  createdAt?: Prisma.DateTimeFilter<"Transaction"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Transaction"> | Date | string
  organization?: Prisma.XOR<Prisma.OrganizationScalarRelationFilter, Prisma.OrganizationWhereInput>
  user?: Prisma.XOR<Prisma.UserProfileNullableScalarRelationFilter, Prisma.UserProfileWhereInput> | null
  bankAccount?: Prisma.XOR<Prisma.BankAccountNullableScalarRelationFilter, Prisma.BankAccountWhereInput> | null
  invoice?: Prisma.XOR<Prisma.InvoiceNullableScalarRelationFilter, Prisma.InvoiceWhereInput> | null
  account?: Prisma.XOR<Prisma.AccountNullableScalarRelationFilter, Prisma.AccountWhereInput> | null
  categoryRef?: Prisma.XOR<Prisma.CategoryNullableScalarRelationFilter, Prisma.CategoryWhereInput> | null
  receipt?: Prisma.XOR<Prisma.ReceiptNullableScalarRelationFilter, Prisma.ReceiptWhereInput> | null
}